OM-54779, RB-37466: Count all entity paths multiple times instead counting an entity only once in the severity breakdown
Intent
======
Count all entity paths multiple times instead counting an entity only once in the severity breakdown so that a critical action on a host shared by multiple applications has greater impact on the severity breakdown.

Requirements
============
1. BusinessApplication, BusinessTransaction, and Service can directly contain Application Component, Database Server, Container, and Virtual Machine
    1. we have to support these relationships because the Settings->ARM Wizard supports them
2. BusinessApplication can contain BusinessTransaction and Service
    1. we have to support these relationships because the Settings->ARM Wizard supports them
3. BusinessTransaction can contain Service
4. BusinessApplication, BusinessTransaction, and Service can have multiple paths to entities deeper in the supply chain like a host. All these paths must be counted. See the bottom for samples.

All potential relationships from BApp, BTxn, and Service from items 1,2,3 are drawn visually below:
```
    BusinessApplication --\    /----------------------------------------------------------
   /        |              |  |    ApplicationComponent---------\--------------------\    \
  /         v              v  |        ^                        v                    v    v
 |  BusinessTransaction -->o--/--------/--\-------------->Container----------->VirtualMachine
 |          |              ^              v                     ^                    ^
  \         V              |       DatabaseServer---------------/--------------------/
    ---> Service ---------/
```

Examples
========

**Example 1**: Services end up using the same host
```
     --> Service1 --> App1 --> VM1 ---
    /                                 \
BTxn1                                  --> Host1
    \                                 /
     --> Service2 --> App2 --> VM2 ---
```
So if Host1 is Critical, then BTxn1 double counts the critical entity host 1
1. critical: 2 entities (Host1 thru Service1 and Host1 again through Service2)
2. normal: 4 entities (App1, App2, VM 1 and VM2)

**Example 2**: Service can have multiple applications or databases:
```
Service1 ----> App1 -> VM1----
          \--> App2 -> VM2--  \
                            \  \
                             ------> Host1
```
Suppose Host1 has a critical action then the breakdown for Service1 is:
1. critical: 2 entities (Host1 thru App1 and Host1 again through App2)
2. normal: 4 entities (App1, App2, VM1 and VM2)

Solution
========
1. Compute the breakdown for all entities that directly connect to BusinessApp, BusinessTxn, and Service. These can be: Application Component, Database Server, Container, and Virtual Machine
2. Compute Service using the direct connections from the repository, using the counts from step 1.
3. Compute Business Transactions using the direct connections from repository, using the counts accumulated from step 1 and 2.
4. Compute Business Applications using the direct connections from repository, using the counts accumulated from step 1, 2 and 3.
5. Remove break downs for Application Component, Database Server, Container, and Virtual Machine because they are not used anywhere. It's a waste to keep them around.

pros: re-uses existing repository APIs
cons: asking repository for `PartialEntity.ReturnType.API` includes information that EntitySeverityCache does not need.
cons: more expensive to compute. now we have 5 requests (previous was 2. one to repo for All BApp,BTxn,Service and a second to SupplyChain for the entities underneath the BApp, BTxn, Service.)
cons: uses more memory because we have to keep around the breakdowns for App, DB, Container, VM until the end of the computation

Alternative Solution
========
1. use WITH_CONNECTIONS, instead of API
    1. con: did not work. connections were always empty even though provider is filled in when you as for API. provider is exactly what i want, but only API fills it in
2. add another return type to repository API
    1. con: more duplicated code filling in details specific to a particular request that only entity severity cache needs
3. do not count duplicates
    1.  cons: does not meet requirements
    2.  pros: computes more efficiently
4. instead of 3 requests to get providers for Services, BTxn, BApp, make one request then order by Service, BTxn, BApp
    1. pros: save 2 requests
    2. cons: requires loading the entire response into memory (because of the sort [see this stackover flow article for more details](https://softwareengineering.stackexchange.com/questions/332487/how-does-sorting-with-java-8-stream-work-under-the-hood)) instead of lazily using it piece by piece

+++ com.vmturbo.action.orchestrator/src/main/java/com/vmturbo/action/orchestrator/rpc/EntitySeverityRpcService.java
@@ -189,7 +189,7 @@ public class EntitySeverityRpcService extends EntitySeverityServiceImplBase {
 
         final Optional<EntitySeverityCache.SeverityCount> severityCountOptional = optionalCache.flatMap(entitySeverityCache -> entitySeverityCache.getSeverityBreakdown(oid));
         if (severityCountOptional.isPresent()) {
-            for (Map.Entry<Severity, Long> entry : severityCountOptional.get().getSeverityCounts()) {
+            for (Map.Entry<Severity, Integer> entry : severityCountOptional.get().getSeverityCounts()) {
                 entitySeverityBuilder.putSeverityBreakdown(entry.getKey().getNumber(), entry.getValue());
             }
         }
